feat: move helper functions to src/utils/helper.js (#2)
