+++ eslint.config.js
@@ -1,5 +1,6 @@
 // eslint.config.js
 import js from "@eslint/js";
+import globals from "globals";
 
 export default [
   js.configs.recommended,
@@ -11,6 +12,9 @@ export default [
     languageOptions: {
       ecmaVersion: "latest",
       sourceType: "module",
+      globals: {
+        ...globals.browser,
+      },
     },
     rules: {
       // Aturan custom Anda
