+++ src/index.js
@@ -1,37 +1,41 @@
+// C:\OSS\FOSS_Issue\move_helper_function\src\index.js
+
+import { showError, clearError, getFieldName } from "./utils/helper.js"; // Đã thêm dòng import này
+
 /**
  * Validates a form identified by the given form ID.
  *
  * @since 2.0.0
  * @param {string} formId - The ID of the form element to validate.
- * 
- *  TODO: Add a second parameter (options) for configuring validation behavior:
- *      - isSubmit: (boolean) whether to auto-submit if valid [default: true]
- *      - withResponse: (function) optional callback to handle validation result
- *      - customRules: (object) additional field-specific validation rules
+ *
+ * TODO: Add a second parameter (options) for configuring validation behavior:
+ * - isSubmit: (boolean) whether to auto-submit if valid [default: true]
+ * - withResponse: (function) optional callback to handle validation result
+ * - customRules: (object) additional field-specific validation rules
  */
 function validateForm(formId) {
   const form = document.getElementById(formId);
   if (!form) return; // Exit if form not found
 
-  form.addEventListener('submit', (event) => {
+  form.addEventListener("submit", (event) => {
     event.preventDefault();
 
     // Clear previous error messages
-    form.querySelectorAll('.error-message').forEach((e) => e.remove());
+    form.querySelectorAll(".error-message").forEach((e) => e.remove());
 
-    const inputs = form.querySelectorAll('input, select, textarea');
+    const inputs = form.querySelectorAll("input, select, textarea");
     let isValid = true;
 
     inputs.forEach((input) => {
       clearError(input);
 
-      if (input.hasAttribute('required') && !input.value.trim()) {
+      if (input.hasAttribute("required") && !input.value.trim()) {
         isValid = false;
-        showError(input, `${getFieldName(input)} must be filled`);
-      } else if (input.type === 'email' && !validateEmail(input.value)) {
+        showError(input, `${getFieldName(input, form)} must be filled`); // Đã cập nhật lời gọi getFieldName
+      } else if (input.type === "email" && !validateEmail(input.value)) {
         isValid = false;
         showError(input, `Invalid email format`);
-      } else if (input.type === 'tel' && !validatePhoneNumber(input.value)) {
+      } else if (input.type === "tel" && !validatePhoneNumber(input.value)) {
         isValid = false;
         showError(input, `Invalid phone number`);
       }
@@ -43,77 +47,30 @@ function validateForm(formId) {
   });
 
   /**
-   * Displays an error message below the input element.
+   * Validate an email address.
    *
-   * @param {HTMLElement} input - The input element.
-   * @param {string} message - The error message to display.
-   * 
-   * TODO: Move to utils/helper.js
+   * @param {string} email
+   * @returns {boolean}
+   *
+   * TODO: Move to separate file for email validation if it'll be.
    */
-  function showError(input, message) {
-    const parent = input.parentElement;
-    const error = parent.querySelector('.error-message');
-    if (error) {
-      error.textContent = message;
-    } else {
-      const errorMessage = document.createElement('div');
-      errorMessage.className = 'error-message';
-      errorMessage.textContent = message;
-      parent.appendChild(errorMessage);
-    }
+  function validateEmail(email) {
+    const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    return re.test(String(email).toLowerCase());
   }
 
   /**
-   * Clears the error message for the input element.
-   * 
-   * TODO: Move to utils/helper.js
-   */
-  function clearError(input) {
-    const parent = input.parentElement;
-    const error = parent.querySelector('.error-message');
-    if (error) error.remove();
-  }
-
-  /**
-   * Returns the field name from label if available, else from id.
+   * Validate a phone number.
    *
-   * @param {HTMLElement} input - The input element.
-   * @returns {string} The formatted field name.
-   * 
-   * TODO: Move to utils/helper.js
+   * @param {string} phone
+   * @returns {boolean}
+   *
+   * TODO: Move to separate file for phone validation if it'll be.
    */
-  function getFieldName(input) {
-    const label = form.querySelector(`label[for="${input.id}"]`);
-    if (label) return label.textContent.trim();
-    if (input.id) return input.id.charAt(0).toUpperCase() + input.id.slice(1);
-    return '';
+  function validatePhoneNumber(phone) {
+    const re = /^[0-9\s()+-]{6,20}$/;
+    return re.test(String(phone));
   }
-}
-
-/**
- * Validate an email address.
- *
- * @param {string} email
- * @returns {boolean}
- * 
- * TODO: Move to separate file for email validation if it'll be.
- */
-function validateEmail(email) {
-  const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  return re.test(String(email).toLowerCase());
-}
-
-/**
- * Validate a phone number.
- *
- * @param {string} phone
- * @returns {boolean}
- * 
- * TODO: Move to separate file for phone validation if it'll be.
- */
-function validatePhoneNumber(phone) {
-  const re = /^[0-9\s()+-]{6,20}$/;
-  return re.test(String(phone));
-}
+} // Dấu ngoặc nhọn đóng này là của hàm validateForm()
 
 export default validateForm;
